refactor sounds
